requirements.txt
google-generativeai
streamlit==1.34.0